requirements.txt
absl-py==2.2.2
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
Brotli @ file:///croot/brotli-split_1736182456865/work
causal-conv1d @ file:///home/cc/%E4%B8%8B%E8%BD%BD/%E7%99%BE%E5%BA%A6%E7%BD%91%E7%9B%98/torch%2Btorchvison%2Btorchaudio%2Bmamba/causal_conv1d-1.4.0%2Bcu118torch2.3cxx11abiFALSE-cp310-cp310-linux_x86_64.whl#sha256=a2433764743bfe4eac6630671ce7dfc7c18c006ff8c1e16359edf6df0a585c8e
certifi @ file:///croot/certifi_1738623731865/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
contourpy==1.3.2
cycler==0.12.1
einops==0.8.1
filelock @ file:///croot/filelock_1700591183607/work
fonttools==4.58.0
fsspec==2025.3.2
fvcore==0.1.5.post20221221
gmpy2 @ file:///croot/gmpy2_1738085463648/work
grpcio==1.71.0
huggingface-hub==0.30.2
idna @ file:///croot/idna_1714398848350/work
iopath==0.1.10
Jinja2 @ file:///croot/jinja2_1741710844255/work
kiwisolver==1.4.8
mamba-ssm @ file:///home/cc/%E4%B8%8B%E8%BD%BD/%E7%99%BE%E5%BA%A6%E7%BD%91%E7%9B%98/torch%2Btorchvison%2Btorchaudio%2Bmamba/mamba_ssm-2.2.2%2Bcu118torch2.3cxx11abiFALSE-cp310-cp310-linux_x86_64.whl#sha256=e746cbceb01bee548e93bf46a78e3c3333520ec1c369f3f2c95a1e51cf821264
Markdown==3.8
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
matplotlib==3.10.3
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///croot/networkx_1737039604450/work
ninja==1.11.1.4
numpy==1.26.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.3.0
pillow @ file:///croot/pillow_1738010226202/work
portalocker==3.1.1
prettytable==3.16.0
protobuf==6.30.2
ptflops==0.7.4
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
pysodmetrics==1.4.2
python-dateutil==2.9.0.post0
pytorch-wavelets==1.3.0
pytz==2025.2
PyWavelets==1.8.0
PyYAML @ file:///croot/pyyaml_1728657952215/work
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
safetensors==0.5.3
scipy==1.15.2
simsimd==6.2.1
six==1.17.0
stringzilla==3.12.5
sympy @ file:///croot/sympy_1738108488918/work
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==3.0.1
thop==0.1.1.post2209072238
timm==1.0.15
tokenizers==0.19.1
torch==2.3.1
torch-dct==0.1.6
torchaudio==2.3.1
torchprofile==0.0.4
torchvision==0.18.1
tqdm==4.67.1
transformers==4.43.3
triton==2.3.1
typing-inspection==0.4.0
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.2
urllib3 @ file:///croot/urllib3_1737133630106/work
wcwidth==0.2.13
Werkzeug==3.1.3
yacs==0.1.8
